Renamed auth.ts to useAuth.ts for consistency.
